Adds SemiCompactObservations (#52)
This changes observation encoding to allow a vector of offsets, rather
than a single offset, when encoding observations. This allows us to do
non-sequential observation encoding. For both existing encoders, this
just means a little more work, since they only use sequential
observation space.

This also adds a SemiCompactObservationEncoder, which uses these new
offsets to allow us to encode same-named observations to the same space.
I.e., we can record "hp" and "ready" to the same location in observation
space for all objects.

My un-tested assumption is that this is a more useful encoding for the
agent. Moreover, it reduces our feature list from 38 to 22. This will
also make it less expensive to give a visible inventory to all
buildings, since it will add ~6 features names instead of adding ~50.

This compiles and trains, both with and without the semi_compact_obs
flag. So far this is only a "it doesn't crash" smoke test.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>