Add converter inventories / queues (#53)
Makes all buildings be Converters, and gives them an inventory.
Converters have a recipe, and based on this they'll convert their
inventory from one type to another.

Right now most conversion parameters are hard coded.

My main testing was with much shorter cooldowns, but it felt satisfying
to be able to pick up and drop off a load of ore.

Follow ups:
* Make more of the conversion configurable.
* Clean up how conversion is triggered, and how inventory is accessed.
* Make put / get work on agents.
* Ensure that tracing works with the new actions.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>