Pull out simulation into its own module (#143)
This pulls out all of the code that is used to do inference (i.e.
simulation) into its own library. This makes it easier to avoid cyclic
imports, because rl should depend on simulation, but simulation doesnt
need RL. We also pulled out wandb into util because while mostly we
should only use wandb during RL, that's not the case right now, so this
was necessary to preserve the fact that sim doesnt depend on RL.

This is all a prelude to introducing a typed SimConfig.

Also, add some backwards-compatability to allow us to load old policies
with the new refactor into a metta-scoped agent module.

---------

Co-authored-by: Jack Heart <jack@softmax.com>