Analyzer fixes and tests (#137)
- Changes to eval_stats_analyzer, in particular db._query is not
private, and we cleaned up the way we check for eval data in the
database for candidate and baseline policies when calculating policy
fitness. Notably, if we are missing data for either candidate or
baseline policies for a particular eval, we just continue. This could
happen during some post-hoc analysis, but shouldn't be a problem during
training, since the candidate will be the current checkpoint and the
baselines will be the previous checkpoints
- Wrote unit tests for the eval_stats_analyzer and added them (along
with jack's unit tests for eval/db.py) to the github workflow
- removed a deprecated unit test (checked with david first)