Fix a test (#101)
I broke test_basic's import tests. This passed locally because of left
over compiled modules. I can't tell if it failed on github when I would
have expected it to -- it's possible it didn't but in a non-blocking
manner, and graphite ignored the non-blocking failure.

Anyway, just doing the basic fix for now.

Co-authored-by: S. Alex Smith <sasmith@stem.ai>