More navigation maps and configs  (#164)
This adds a number of new maps and configurations being used to try and
improve agent navigation. The main thrust of this PR is to add harder
evals and make sure all of the eval scores are summing to 1.

Testing:
Ran training with these maps on a Mettabox and AWS

---------

Co-authored-by: daphnedemekas <daphnedemekas@gmail.com>
Co-authored-by: George Deane <georgedeane@Georges-MacBook-Air.local>
Co-authored-by: Jack Heart <jack@loopflow.studio>
Co-authored-by: Daphne <45669133+daphnedemekas@users.noreply.github.com>
Co-authored-by: George Deane <georgedeane@Mac.localdomain>
Co-authored-by: Jack Heart <jack@Jacks-MBP.localdomain>
Co-authored-by: Jack Heart <jack@stem.ai>
Co-authored-by: Jack Heart <jack@softmax.com>
Co-authored-by: George Deane <georgedeane@MacBook-Pro.local>
Co-authored-by: Lars Sandved <lars.sandvedsmith@gmail.com>
Co-authored-by: George Deane <george@stema.ai>