fix: link workspace to conda via symlink (#161)
As long as the path is referenced at an absolute location it seems that
we can use a symlink in the VS Code workspace.

> [!NOTE]
> This requires that developers add a symlink to tmp after setting up
conda:
> ```sh
> conda activate metta
> ln -s "$CONDA_PREFIX/lib/python3.11/site-packages"
/tmp/conda-site-packages
> ```

This PR updates the workspace settings to include
`/tmp/conda-site-packages` in the Python analysis paths, which helps VS
Code resolve packages from Conda environments correctly.