Fix the eval plots (#173)
Since I merged in the NPC changes, we use the given title in the eval
config as the eval name, which breaks the way that the trainer selects
whether the eval is navigation, object_use or npc. So here, I just amend
the eval names in the config to account for that.