Av-ls-ag-eigen (#147)
This branch contains the code for plotting various metrics during
training that can be used to infer the degree of criticality for each
layer.

The metrics are calculated using the analyze_sv method in dynamics.py
and include:

- mean singular value -- This is related to the branching coefficient
and is near 1 at criticality
- largest singular value
- ratio of max_sv / min_sv ("the condition number") 
- the R^2 of a linear fit in loglog space of the singular value
distribution -- good fit is >0.95
- the slope of that fit -- in range [-1.5,-.5] at criticality
- the % of singular values that have collapsed to near zero -- gives
information about the appropriateness of the layer size

These are calculated for each layer individually, currently plotting for
actor_1, critic_1, fc1, and encoded_obs.

Agent configs should have the argument "effective_rank: true" for every
layer of interest. Metrics are caluclated every effective_rank_interval
timesteps.

---------

Co-authored-by: al-does <avardakostas@gmail.com>
Co-authored-by: Alex Vardakostas <134578379+Al-does@users.noreply.github.com>
Co-authored-by: David Bloomin <daveey@gmail.com>