Add a typed SimulationConfig (#156)
This change introduces our first dataclass config object, and attempts
to add some patterns that can be copied to convert the rest of the
codebase. The high level idea is that we use Hydra just for converting
from a dir of yaml files + cmd line args into a single config dict, then
we convert that using our own logic.

In particular:
* each tool has its own top-level config (defined in the cmd.py file
itself) that can be parsed into a typed dataclass made out of component
configs. We can use both these and the regular dictconfig until all
config types have been converted.

* there is a preprocess hook that allows you to do thing like propagate
fields to children for specific config types. In particular, this allows
the SimulationSuiteConfig to define defaults for all its
SimulationConfigs

A few related changes:
* a few more things were pulled into metta/sim, so that all the
simulation style code (for stats, human play, and log replay) can live
there and use a shared SimulationConfig (TODO: actually convert them to
all using the same Simulation class)
* attempted to leave the yaml files mostly the same. I considered moving
all the job yamls into a separate dir, but that gave me some issues, so
i left them at the top level for now

Design Notes:
I played around with using a variety of approaches for how we do the
dictconfig_to_dataclass, such as `OmegaConf.structured`. However, I
settled on this approach because it makes sure we end up with actual
dataclasses (not DictConfigs that implement a dataclass-interface, which
is what structured does). I introduced the propagate_cfg logic because I
felt it was best to mutate the dictconfigs for custom default-adding
logic, both because hydra already does a lot of that and also because
this leaves our dataclass instantiation and validation very clean. I
tried doing something with __post_init__ on the dataclass objs, but the
problem is that you cant actually instantiate invalid-dataclasses
(good!) so you you cant mutate *them* to add defaults.

Testing:
* added unit tests for SimulationConfig
* ran all the various commands (play, replay, train, sim, analyze)