Fix for metrics boolean attribute error (#195)
Just a simple fix for the error Daphne highlighted. 

The problem was that previously trained models do not have the boolean
flag as a component attribute. The fix is just adding a check for the
presence of that attribute and returning None in its absence.

Also changed the default compute interval in simple.yaml back to 300.

Co-authored-by: David Bloomin <daveey@gmail.com>