Add offset to mapgen Node.select_areas. Add tests (#240)
Adding an `offset` parameter allows more precise specification of which
area of the node you're trying to select. Depend on how it's used, this
can be redundant with `tags` -- I'll have another diff for that shortly
-- but for now this more readily allows for things like "lock the 2nd
and 4th area so you don't put altars in them".

This also adds unit tests, partially to carve out more of an expectation
for there to be tests.