Add layouts to room_grid (#241)
Adds the option to provide a layout to RoomGrid, which will create a
room in which areas have a given tag. Expands the automatically-added
tags to include "room_x_y", since cursor auto-completed to doing this,
and it seemed like a good idea.

This adds layouts to RoomGrid instead of creating a RoomLayout since
when I made the latter, the code seemed overly duplicative.

Adds tests.