Add left vs right map, where agents need to choose a direction. (#252)
This is a map where agents need to decide to go to the left or right to
get reward -- not enough time to do both.

This is intended as a basic env for diversity training / eval.

This doesn't currently track whether the reward was placed on the left
or right.