Remove python functions from ObservationEncoder (#259)
This removes certain Python functions from ObservationEncoder, making it
more possible for it to be fully cpp.

In both cases, the functions blurred the line between the grid and
observations. Since I wanted to move the ObservationEncoder into cpp, I
pushed these into the grid as needed, but long term we'll need to
develop a clearer divide.

* _obs_encoder.observation_space isn't meaningfully used -- things just
use the grid's understanding of observation width and height instead.
* _obs_encoder.obs_np_type was used, but it was constant and it's hard
coded for now. At some point I think we'll care about the the data types
we're using, and it seems likely we solve that with templating.

Tested via compilation.

Co-authored-by: S. Alex Smith <sasmith@stem.ai>