Convert observation_encoder to cpp (#260)
Converts ObservationEncoder to cpp, including making is a cpp class
instead of a Python class. This included a small amount of pushing on
boundaries (e.g., where is the ObservationEncoder created), but all of
the boundaries felt sort of arbitrarily placed, and so it seemed
reasonable to shift them in order to let this be a fully cpp class.

Tested by compiling and smoke-test training on my laptop.