EFS proxy opentofu configs (#261)
# Tailscale

This PR (`tailscale_acl.acl` resource in `tailscale.tf`) **replaces the
Tailscale ACL**. Be careful when applying configs, you probably will
need to update `tailscale.tf` first.

# OpenTofu vs Terraform
I've used OpenTofu but I expect both OpenTofu and Terraform would work
identically.

The lockfile is generated with OpenTofu though, and Terraform might try
to reformat it.

(My impression is that OpenTofu is doing well, but I haven't followed
the fork story closely in the last year.)

# Performance

Latency is high. Copying a single 300KB file to the local share takes 12
seconds for me. OTOH, it's fast when I mount it on EC2, so it might be
better for others.

# Installing on macs

Not automated yet; this PR has Terraform configs for AWS. I'll add a
script for macs soon.

---------

Co-authored-by: David Bloomin <daveey@gmail.com>