Merge mettagrid and grid_env (#303)
Merges MettaGrid into GridEnv, since we're not longer trying to support
GridEnv as a generic thing on which we'll build a bunch of different
environments.

The merge is mostly a copy / paste, and doesn't clean things up as much
as I wish it did, but as per some of the comments, I don't want to break
up existing Python functions into a bunch of smaller Python functions at
the moment, since the intention is to convert away from Cython anyways.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>