update initial torch sync epoch (#389)
Single line change: increased the first profile epoch from 50 to 300 in
the torch profiler. At 50, it would annoyingly stop training for people
where trying to run locally or while actively remoted into a console for
testing. 300 is when we sync with wandb and run sim anyways so it seemed
like a good place to slot in the first trace. Ideally, the first trace
should also run early enough that perf issues can be debugged without
having to restart the run.