MultiLeftOrRight scene; order_by_seed feature (#393)
As requested by Lars.

I had to copy-paste left_or_right yaml config to Python. The better
solution would be to move it to a scene and reuse, but scenes root is
currently in mettagrid, so can't do that right now.

This PR creates a precedent of `metta.map.scenes.*` files; I'd argue
that we should move all mapgen code to `metta.map.*` and leave mettagrid
minimal/agnostic to map builder implementations.