Make the arguments for TerrainFromNumpy the same as Random (#417)
Also moved the file into the main section of mettagrid environments. 
Only difference is that you can't pass in a width and height (since they
are preloaded)

Example replay:
https://metta-ai.github.io/metta/?replayUrl=https%3A%2F%2Fsoftmax-public.s3.us-east-1.amazonaws.com%2Freplays%2Flocal%2Fdaphne.local.4%2Fterrain_from_numpy%2Freplay.json.z&mapPanX=-6849&mapPanY=-23246&mapZoom=0.025

The file names of the numpy arrays do have widthxheight in them, so if
we wanted to constrain by width and height which maps we select, that
could be done.