Add back an expanded assertion (#462)
This adds back an assertion about the type of `metta_agent`.

This assertion was added by @rjwalters to add code clarity. It was
originally added to make the type information we were showing -- that
metta_agent was a `MettaAgent`. This turned out to be wrong, since it
could also be a `DistributedMettaAgent`.

I don't have a strong opinion on the value of the assertion; but I'd
fixed it this way before the assertion was commented out (or at least,
before that change got to my branch), so I figured I'd propose this fix.

Co-authored-by: S. Alex Smith <sasmith@stem.ai>