eval smoke test (#517)
Add a smoke test for evals.

This refactors the run_evals script and sim.py to allow us to indicate
we're running a smoke test. The smoke test runs on a much stripped down
number of evals and simulations.

This also updates the policies we're using in run_evals to match
https://github.com/Metta-AI/metta/blob/daphne-evaluation/run_evals.sh.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>