feat: add workflow for launching AWS training jobs (#573)
This PR adds a new flag to our AWS launch script that allows a job to
self-terminate after a certain number of minutes.

The intention is to use this in a new workflow that will train for ~30
min on env/simple for every push to main so that we can spot real world
performance problems earlier.