Switch away from using a numpy array for maps (#551)
This moves us from using numpy arrays for maps at the cpp level to using python lists. I think there are better solutions -- ones where we reconsider using numpy arrays in Python as well, and maybe rethink using strings as values, but this is easier to apply surgically.

The main reason for this change is to be able to switch to using normal Python _strings_, so we can avoid needing to do utf32 -> utf8 conversion in cpp, and can get rid of a lot of warnings about the deprecated libraries we're using to do so.