fix: don't retry jobs when we abort on SIGTERM (#584)
The AWS jobs we are launching are quitting after 60 minutes but then
restarting because of our AWS settings. This PR applies two separate
fixes -- first we convert SIGTERM to exit 0 which should make AWS think
the job has completed normally and the second is to limit retries in
case this doesn't work.