Update inventory management (#599)
Refactors inventory management to make objects more in control of when they can and can't receive input.

This changes to make it so put and get return the actual delta of items, rather than having outside logic assume what the agent / converter can take. The main purpose of this is to make it so we can [later] have per-item inventory limits, and have the logic to manage those contained in the `update_inventory` function. As a side effect of this change, I fixed some places where items could get lost, like when you're stealing from a frozen agent.

This still makes some assumptions that entities will be able to _lose_ inventory if they have it. If we cared, we could also expose functions like "max_potential_update", rather than only exposing it through mutation.

Testing:
Added unit tests.