remove mettagrid make call in setup_build.sh (#682)
Earlier versions of #630 included `make` call in `setup_build.sh`, so
that devs could run tests.

Then I removed makefile but forgot to update it.

Instead of running cmake commands, I decided to skip it entirely - which
means that to run mettagrid C++ tests or benchmarks locally we'll have
to run `cmake` manually (as explained in mettagrid README).