Move eval dashboard to react (#641)
This PR does a few things
1) Moves the eval dashboard to react
2) Moves the eval dashboard model to something resembling replays, where
the python code generates data, and then that data is rendered by the
javascript / react code to display the dashboard. Currently that data is
a json file that is generated by the 'dashboard' tool. It can either be
passed in as a url parameter (if it is hosted on s3), or uploaded from
local. When we choose to, we can swap that out for a hosted database.
3) A couple of small dashboard improvements, like supporting multiple
eval suites and metrics in the same dashboard

The latest version of the dashboard is deployed at
https://softmax-public.s3.amazonaws.com/observatory/index.html - you can
preview it with some data at
https://softmax-public.s3.amazonaws.com/observatory/index.html?data=https://softmax-public.s3.amazonaws.com/users/ptsier/pasha_data_4.json

Generating the data is done, as before, using the 'dashboard' tool - for
example, the data for the above was generated using the command
`python3 -m tools.dashboard +eval_db_uri=wandb://stats/pasha_combined_2
run=pashadashboard
++dashboard.output_path=s3://softmax-public/users/ptsier/pasha_data_4.json`

---------

Co-authored-by: Paul Tsier <paul@elementanalytics.com>