Diversity bonus reward mechanism and evaluation environment (#518)
This PR is for the diversity reward bonus mechanism, which scales agent
rewards at the end of the episode - increasing more if the agents reward
is close to their own group mean and far from other groups.

**Implementation details:**

- Turning on the bonus is achieved by adding a `enabled: true` flag to
your environment .yaml under game>diversity_bonus. This flag is default
`false` in the base configs. You can also adjust the coefficients that
scale the similarity and diversity pressure (currently default = 0.5) by
adding e.g. `similarity_coef: 10` and `diversity_coef: 10`. For example:
```
game:
   diversity_bonus:
      enabled: true
      similarity_coef: 10
      diversity_coef: 10
```
- When enabled, the bonus is applied at the end of the episode within
the mettagrid_env step function. The bonus applied multiplicatively to
increase the reward for within group similarity and between group
difference.

- The `calculate_diversity_reward` function in
`mettagrid/util/diversity.py` takes the agent's end of episode reward
and computes the normalised absolute distance between their reward and
their own group's mean reward, as well as the normalised absolute
distance to other group rewards. This distance is converted into a value
between `[0,1]`, such that the scaling factor is close to 1 for
similarity within group and difference between groups, and close to 0 in
the opposite case. These factors are multiplied by the coefficient
parameters and summed to form the reward bonus scaling.

**Evaluating diversity pressure**
- The `multi_left_right.yaml` environment is designed to test the
effects of the diversity reward mechanism. In this map agents find
themselves in the middle of a rectangular room with only enough time to
explore one side (25 timesteps). Each map generates 64 rooms with an
agent in each, randomly allocated to `agent_group: 1` or `agent_group:
2`.

- At every episode the rooms are all identical, with reward randomly
placed on either the left or the right (2 single use altars).

- The hypothesis is that the diversity reward should drive the agents to
adopt a diversity of strategies, in this case we would expect to see one
group adopt a 'go left' and the other adopt a 'go right' strategy.

- So far this is not what I'm seeing. I think the issue is that the
'wrong' strategy gives exactly zero reward, which means that the agents
have no increased incentive to do that (since the bonus has no effect).
I tried adapting this to be additive, where the bonus is added to the
reward at the end of the episode. But this led the agents to not learn
anything since they could happily do nothing and receive reward. So the
next experiment I'll run is to give a different reward in each direction
(e.g. 1 altar left and 3 altars right) and see if we get the expected
effect.

- More sweeps across the coefficients are also required. 

There are many ways to make the reward mechanism more sophisticated when
we choose to, e.g.:
- have species be an embedding rather than a discrete value
- separate group and species, so that we can have teams comprised of
different species
- make the calculation signed such that there's a penalty for being
different to your group and same as the others. Currently the bonus is
always positive
- calculate the distance based on the reward vector across the episode
instead of the end of episode reward. This would bring strategies closer
together, instead of just final reward.
- etc