feat: Have summary of PRs send to Discord channel every week (#741)
Sends a Discord message to the `#weekly-pr-update` channel every week on
Friday at noon that summarizes the PRs that have been merged in the last
seven days.

The prompt could certainly be a little more optimized to our
preferences; feel free to make changes.

**Changes**
Add a new GitHub workflow to do that.

**How?**
1. Retrieve all PRs from GitHub that have been merged in the last seven
days, then also retrieve a truncated version of their diff and PR
description.
2. Give all that to an LLM.
3. Send the summary to the Discord channel using a webhook.

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>