Move normalization values to the mettagrid environment (#794)
This moves the source of truth for observation normalization values from the ObservationNormalizer to the cpp environment. This sets us up to have the normalization set in configuration.

We continue passing grid_features for now, although they're unused. We should talk about whether they're useful to ensure that the features remain stable (e.g., the order remains stable). This is a similar problem we have for tokens -- making sure the feature_ids remain appropriately stable.

Normalization values are also only passed on layer creation -- not when we save and restore. That will need to change if we want the agent to be able to deal with normalization values updating between environments.