Bring mettamap to monorepo; preview mettagrid configs with it (#784)
## How to run

```
# in one terminal:
python -m metta.map.server

# in another:
cd mettamap
pnpm install
pnpm dev
```

## History

This code is based on what I had previously in mettamap repo's
[player](https://github.com/Metta-AI/mettamap/tree/player) branch. That
branch had an experimental frontend for `tools.play` that I did a month
ago, but now MettaScope has this functionality so I removed that code.
But the map viewer functionality (with pan & zoom) remains.

I've also added some new features.

## Configs viewer

![Screenshot 2025-06-03 at
1.07.50 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/evCtXUW5JNyNfqWDHydW/a04cf3af-9b43-4495-ab3f-13584ceb7510.png)

![Screenshot 2025-06-03 at
1.09.11 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/evCtXUW5JNyNfqWDHydW/1f6e63a4-9f28-4c13-a982-3c917de5984a.png)

Non-obvious feature: config name and class names can be clicked and
they're immediately opened in local Cursor editor.

## Viewing stored maps from S3

We're not currently using the ability to store mapgen maps, I'm not sure
if we will, but I kept the code and switched it from local files to S3,
as we discussed before.

## Other notes

This is still unpolished and I know there's a lot that could be cleaned
up, but this PR is pretty big as it is, so I want to send it in early.

Some especially unpolished things:
1. `metta.map.server` could be merged with `mettascope.server`
2. Code is separate from the observatory
3. "Stored Maps" is very slow (because of S3)
4. "Stored Maps" part hardcodes a single map collection on the backend,
`test-collection`
5. Pan&zoom cell highlighting is a bit broken
6. Maps with blocks don't render (not an issue with this viewer, though,
it's a mapgen StorableMap limitation that I'll fix separately)
7. Maps use old style sprites.

I hope to fix these soon; (1) and (2) require coordination with Pasha
and Andre; (3)-(7) could be improved but I had to stop and submit
somewhere. I can continue to fix in this PR or in follow-up PRs.