Assert that we're not copying `actions` when receiving them in mettagrid (#835)
Update the MettaGrid C++ interface to explicitly use `int32_t`
parameters instead of the generic `int` type, and assert that we're not
making unexpected copies. This should _not_ be a correctness issue,
since we only read from the actions array; but we've gotten burned by
unexpected copying before, so it's nice to root out.

Tested by running `train`. If this were incompatible with what puffer
was sending, I would expect to see an immediate crash.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>
Co-authored-by: Robb Walters <rjwalters@users.noreply.github.com>
Co-authored-by: Robb Walters <robb@2amlogic.com>