Switch normalization to be a map (#970)
Normalization was originally created as a list to match feature_names
(which we don't export anymore). I thought a list made sense, since we
could have some expectation that the feature_id space could be dense.
I.e., when initializing, we would re-map whatever ids we had to be
dense.

I'm back-tracking on this now. I think it's still possible, but it feels
like an unnatural representation of the data we have. This became more
apparent when trying to add normalization for global features -- it felt
like keeping the list (/ vector) structure was unnecessarily burdensome.

This switches all of our policies to be able to use the dict / map
version as well; but I'm not clear on what the expectations are around
backwards compatibility. I _think_ there's nothing to do, since the
models were pickled with the normalization they had when they were
created.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>