get rid of feature_names() (#972)
Generate offsets directly using feature_ids, instead of implicitly based
on feature_name. This lets us remove feature_names.

An important change here is that this also switches normalizations to be
by feature_id as well. Previously they were generated based on the order
in which feature_names were encountered. This was related to, but not
that same as, the order given by feature_id.

I think the normalization _happened_ to be right, prior to this change
(and global features); but it was mild accidental, and definitely
fragile.

Note that I'm trying to fix this for existing box observations rather
than ripping out box observations, since I want us to have something
correct to fall back to.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1210550931090512

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>