homogenize how we add tokens (#978)
Standardize how observation tokens are added, so we can add global
observation tokens safely.

This comes from hitting some tests where we didn't specify the number of
tokens and hit errors. We should still have warnings about having too
few tokens in most circumstances; but regardless it's more correct to
not try to write past the end of the vector.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>