Puffer3 Trainer (#839)
Hi,

This is an overhaul of `trainer.py` and `experience.py`. The new trainer
and experience are lifted from the current PufferLib/ implementation in
(https://github.com/PufferAI/PufferLib/blob/3.0/pufferlib/pufferl.py)

The changes include a re-work of a number of variables as well as the
inclusion of two new features:
- advantage filtering
- prioritized experience replay-ish

Tomorrow I will test that these changes work on skypilot as the
mettaboxes are down and I built the new 3.0 image there and need to push
it, so will push to ECR tomorrow morning.

But for now looking for comments so I can make necessary changes. But
before we merge this, the things we still have to do include:
- push container to ECR tomorrow
- debug multi-gpu on skypilot
- run object use evals on skypilot
- run nav sequence evals on skypilot
- run memory evals on skypilot

Already on a mettabox this branch is number one on our nav leaderboards!


![image](https://github.com/user-attachments/assets/d65bf9a1-768a-42b3-92fe-ecd09ffb9584)

Overall, I tried to keep the diff small where able, with the flowing
mostly being similar, including keeping a separate experience class
still.

---

Other minor changes include:
- updated puffer.yaml to match mettagrid.yaml, will need to change this
to be explicit everywhere for joseph's ease of use later.
- added v-trace parameters to the config, though I think the current
pufferlib trainer isn't actually using the corrected importance sampling
ratios
- asserts for wandb=off from our outage
- currently pointing to a commit on my pufferlib branch because it
included a fix for a memory leak and I want to see if this is why the
mettaboxes crash after running for many hours
- most other functionality remains the same, it is TBD to debugging
distributed for having skypilot have my image tomorrow morning