Self destroy jobs that got stuck and inform user (#957)
**Overview**
We modify the heartbeat system introduced in #661 to require manual
calls to the heartbeat recording function.
If no heartbeat has been detected for a certain amount of time,
`wandb.alert()` is called which notifies the user of this timeout before
killing the training process.

**Technical internals**
Allowing the `wandb.alert()` to be send was quite challenging as the
heartbeat monitor lives in another process and can't just access the
same wandb run object. To solve this problem we use IPC over file to
share the wandb run_id and other internal with the monitor process.

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>