Performant token obs encoder (#1011)
Revised to be performant both in sample efficiency and SPS.

### Refactors
- Removed legacy and experimental observation encoder configs and code.
- Added new obs_tokenizers.py with modular tokenization and
normalization layers for better composability.
- Significantly improving SPS by implementing a GPU-efficient way of
stripping padding beyond the longest dense sequence in a batch.
- Updated agent configs to use the new encoder and increased the number
of observation tokens for the environment so we don't drop tokens.
**`configs/agent/simple.yaml` now uses the tokenized encoder.**
- Enhanced self and latent attention blocks to be fully functional,
supporting multiple heads, multiple layers, and the optional use of a
classifier token. If the latter option is used then it can output
directly to the LSTM. However, this means that it cannot output directly
to another attention block.

### Testing
- `simple.yaml` performed close to the original CNN and
`latent_attn_med.yaml` performs better than the CNN when trained on
navigation curricula, according to navigation evals.
- Extensively tested SPS on a mettabox. The original CNN runs at roughly
200 SPS, 'simple' runs at 180 SPS, and 'latent_attn_med' runs at 110
SPS. Note that these numbers for the original CNN are lower than they
were before puffer 3.0 transition as the sheer size of the refactor
means that, inevitably, not all features can be ported over at once.
Once those features return and the original CNN's speed returns closer
to where it was (400 SPS), we should test the token obs encoder models'
speed again.