[Github-Asana] Get the right users (#1069)
Previously we weren't filtering which users we were getting back from
the roster. This might be fine, except that we abort when we get as many
users as we're asking for. So we were getting the wrong ones!

This worked fine during initial testing, since I hadn't populated the
roster. Now that I have, this bug matters.

I should probably right a test, but it's a _little_ hard because this is
a github workflow, so I'm going to delay that for now.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210597267013782)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>