[MettagridConfig] Enforce strict config validation and add max ore/battery fields (#1084)
This PR introduces a base model with `extra="forbid"` to enforce strict
validation across all config classes, and adds fields required to make
this possible.

I'd previously thought that the `extra="forbid"` was effectively
inherited from `GameConfig` to all of it's fields; but this appears not
to be the case.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210620117482737)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>