Refactor 'Launch SkyPilot Training Job' to use reusable GitHub action (#1105)
The GitHub workflow 'Launch SkyPilot Training Job' handles all the
necessary credential stuff to simply create a new SkyPilot training run.
So far, this workflow is being used for a smoke test on new merges to
main.

**Changes**
Refactor the 'Launch SkyPilot Training Job' workflow to use a newly
created GitHub action to deploy the SkyPilot job.

**Context**
I am working on a smoke test for distributed training that also needs to
deploy SkyPilot jobs, and copying this workflow would be suboptimal.