Adding smoke test for distributed training (#1143)
We want to verify in regular intervals that using 4 GPUs instead of one
results in the same performance per global agent step and an approximate
4x of reward over time.
This PR is the first step towards such a system.

**Changes**
Add a GitHub workflow that deploys two SkyPilot jobs every Sunday (1
GPU, 4 GPUs).

The reason I already create the workflow and only later on add more
functionality on top of this is mainly that the workflow has to already
exist in main in order for me to test later iterations in local
branches.