Remove converter type_ids from cpp (#1196)
Adds explicit type_id fields to all object configurations in the
MettagGrid environment. This change decouples the object type
identification from hardcoded enums.

Right now the type_ids are explicit in config files, which seems
probably not where we want to end; but we need them for the moment
because we want the type_ids to be stable. Happily, the iteration on
getting rid of this should now let most of this iteration be moved out
of cpp.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210654310008279)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>