Fewer default train steps: from 50b down to 10b (#1302)
This is to prevent long, expensive runs that should have been killed and
where the experiment was not intended to go as long. 10b is plenty of
time for a lot of the dynamics we might test for given the current
configs. The intended behavior is for the experimenter to be intentional
about exceeding 10b.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210704144034071)

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210704439703012)

Co-authored-by: Richard Higgins <richard@relh.net>