Expose cpp config objects in Python (#1310)
Refactors object configuration to use C++ classes instead of Python
dictionaries, at the bottom of the config tree.

This PR introduces a new `GridObjectConfig` base class and refactors the
configuration system for grid objects to use proper C++ classes instead
of Python dictionaries. The changes include:

- Added a new `GridObjectConfig` base class that all object configs
inherit from
- Exposed AgentConfig, etc to Python via pybind
- Use dynamic casting rather than `object_type` to figure out the type
of the object, and deprecate `object_type`.

There's a lot of verbosity here that feels redundant: the constructors
that just copy all the arguments into properties, the inits that do the
same, the exposing of every property manually. But as far as I know,
every layer is needed. So here we are.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210712662045209)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>