Correctly zero out zero_actions (#1335)
Properly initializes the `zero_actions` array in the `reset()` method by
filling it with zeros. Previously, the array was created but not
initialized with values. The PR also improves test error messages by
adding more detailed context when assertions fail.

Note that the test in question tended to pass when the fully test suite
was run, but fail when the test was run individually. I assume this is
because when the full test suite was run, memory happened to get
cleared.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210721549334192)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>