refactor agent config (#1338)
This is a partial refactor / rearrangement of agent config. This:

* Pushes item limits under a `resource_limits` setting, letting us drop
the `_max` suffix, and letting us stop polluting the namespace in a
weird way.
* Removes the hardcoding of expected names for resource_limits.
* Right now we won't catch if people set limits on things that don't
exist.
* Moves `invalid_action_penalty` out of `rewards`, since it doesn't
match everything else there. Also, deletes it from all configs since
it's not being used.
* I do worry this is deceptive, since if people want to set it, they'll
most naturally set it in the wrong place. Hopefully the error messages
will be clear enough for them to figure it out!
* Renames some amount of "items" to "resources".

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210722120063864)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>