validate curriculums; fix more env configs (#1343)
Continuing the approach from #1331: validate all curriculum configs.

In current `main` we have 22 curriculum configs, 1 is partial, 7 are
valid and 14 are broken.

This PR fixes all of them.

I realize that some of these configs will go away after Daphne's #1155
and #1093, and it will make the merge of those PRs slightly more
complicated, but I still think it's a good idea to get this in now (in
light of ongoing mapgen updates, I need to test for regressions)


[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210724863740215)

---------

Co-authored-by: David Bloomin <daveey@gmail.com>