More mapgen-powered envs (#1341)
This PR:
- converts all ascii rooms to mapgen (and added `MapgenAscii` subclass
for less verbose configs)
- integrates features from `room.LabyrinthMaze` into `scene.MazeKruskal`
and renames it to `scene.Maze` with `algorithm` parameter
- ports `RadialMaze` to mapgen
- ports `MeanDistance` to mapgen

All updated envs should be identical or almost identical to their
previous versions.

In the code ported from rooms, I sometimes removed broken features, or
removed the features that could be reimplemented with mapgen's
sub-scenes in configs. For example, `LabyrinthMaze` was placing 3 altars
as the reward in the middle of the maze; now it's done by drawing a
custom `InlineAscii` on top of the labyrinth.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210723674896096)