Fixed error initializing teacher policy (#1380)
Teacher previously attempted to be initialized with None which caused an
error in MettaAgent at line 200 when it tried to grab features.items().
Changed it to give teacher the same env features that the student sees.
It's expected that the teacher's previous mapping is saved uniquely to
the object and remapped accordingly.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210747988595233)

---------

Co-authored-by: Richard Higgins <richard@relh.net>
Co-authored-by: Richard Higgins <richard@softmax.com>